refactor: 관리자 페이지 코드 리팩토링
