+++ FE/src/types/authAdmin.ts
@@ -1,0 +1,6 @@
+export interface AuthAdmin {
+  email: string;
+  nickname: string;
+  password: string;
+  oAuthId: string;
+}
